Adding encoding utils for decode_base64url function, small syntax fixes. Adding encoding utils for decode_base64url function, small syntax fixes.
(#86c779jzg)
