+++ norman_utils_external/key_utils.py
@@ -6,6 +6,9 @@ from cryptography.hazmat.backends import default_backend
 from cryptography.hazmat.primitives import serialization
 from cryptography.hazmat.primitives.asymmetric.rsa import RSAPublicNumbers
 
+from norman_utils_external.encoding_utils import EncodingUtils
+
+
 class KeyUtils:
     @staticmethod
     def jwks_to_public_key(jwks: list[dict]):
@@ -14,24 +17,24 @@ class KeyUtils:
         modulus_bytes = jwk.get("n")
         exponent_bytes = jwk.get("e")
 
-        modulus = int.from_bytes(KeyUtils.decode_base64url(modulus_bytes), byteorder="big")
-        exponent = int.from_bytes(KeyUtils.decode_base64url(exponent_bytes), byteorder="big")
+        modulus = int.from_bytes(EncodingUtils.decode_base64url(modulus_bytes), byteorder="big")
+        exponent = int.from_bytes(EncodingUtils.decode_base64url(exponent_bytes), byteorder="big")
 
         public_numbers = RSAPublicNumbers(exponent, modulus)
         public_key = public_numbers.public_key(default_backend())
 
-        public_key_pem_bytes = public_key.public_bytes(
+        public_key_bytes = public_key.public_bytes(
             encoding=serialization.Encoding.PEM,
             format=serialization.PublicFormat.SubjectPublicKeyInfo
         )
-        pem_public_key = public_key_pem_bytes.decode("utf-8")
 
-        return pem_public_key
+        decoded_public_key = public_key_bytes.decode("utf-8")
+        return decoded_public_key
 
     @staticmethod
-    def public_key_to_jwk_list(public_key_pem: str):
-        public_key = serialization.load_pem_public_key(public_key_pem.encode())
-        public_numbers = public_key.public_numbers()
+    def public_key_to_jwk_list(public_public: str):
+        encoded_public_key = serialization.load_pem_public_key(public_public.encode())
+        public_numbers = encoded_public_key.public_numbers()
 
         modulus = public_numbers.n
         exponent = public_numbers.e
@@ -63,10 +66,3 @@ class KeyUtils:
 
         jwk_list = [jwk]
         return jwk_list
-
-    @staticmethod
-    def decode_base64url(data: str):
-        padding = 4 - len(data) % 4
-        if padding != 4:
-            data += '=' * padding
-        return base64.urlsafe_b64decode(data)
